pyproject.toml
[project]
name = "exp-env"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "yama-yeah", email = "b1021091@fun.ac.jp" }
]
dependencies = [
    "ipykernel>=6.29.4",
    "pandas>=2.2.2",
    "polars>=0.20.18",
    "pyarrow>=15.0.2",
    "matplotlib>=3.8.4",
    "torch>=2.2.0",
    "scikit-learn>=1.4.2",
    "sentence-transformers>=2.6.1",
    "transformers>=4.39.1",
    "setuptools>=69.3.0",
    "gensim>=3.3.2",
    "scipy==1.9.3",
    "zetascale>=2.4.2",
    "torchsummary>=1.5.1",
    "fairscale>=0.4.13",
    "fire>=0.6.0",
    "tiktoken==0.4.0",
    "blobfile>=2.1.1",
    "better-loggers>=1.0.1",
    "adabelief-pytorch>=0.2.1",
    "torcheval>=0.0.7",
    "seaborn>=0.13.2",
    "plotly-express>=0.4.1",
    "nbformat>=5.10.4",
    "cellpylib>=2.4.0",
    "sentencepiece>=0.2.0",
    "protobuf>=5.27.1",
    "wandb>=0.17.3",
    "nbconvert>=7.16.4",
    "swifter>=1.4.0",
    "memory-profiler>=0.61.0",
    "cupy-cuda12x>=13.2.0",
    "shap>=0.46.0",
    "japanize-matplotlib>=1.1.3",
    "signalz>=0.8",
    "datasets>=3.2.0",
    "umap-learn>=0.5.7",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
hello = "exp_env:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/exp_env"]
